Use widgets for forms instead of functions
