+++ lib/views/auth/auth_page_form.dart
@@ -1,11 +1,7 @@
-import 'package:client_common/api/response_models/api_error.dart';
 import 'package:client_common/models/auth_model.dart';
-import 'package:client_common/navigator/common_navigator.dart';
-import 'package:client_common/views/error.dart';
-import 'package:client_common/views/loading_button.dart';
-import 'package:flutter/gestures.dart';
+import 'package:client_common/views/auth/login_form.dart';
+import 'package:client_common/views/auth/register_form.dart';
 import 'package:flutter/material.dart';
-import 'package:go_router/go_router.dart';
 import 'package:lenra_components/component/lenra_text.dart';
 import 'package:lenra_components/lenra_components.dart';
 import 'package:logging/logging.dart';
@@ -21,16 +17,9 @@ class AuthPageForm extends StatefulWidget {
 
 class _AuthPageFormState extends State<AuthPageForm> {
   final logger = Logger('AuthPageForm');
-  final _formKey = GlobalKey<FormState>();
 
   bool isRegisterPage = true;
-  String email = "";
-  String password = "";
-
-  bool keep = false;
-  bool _hidePassword = true;
   var themeData = LenraThemeData();
-  bool isLogging = false;
 
   String? redirectTo;
   String? appServiceName;
@@ -43,13 +32,6 @@ class _AuthPageFormState extends State<AuthPageForm> {
 
   @override
   Widget build(BuildContext context) {
-    isLogging = context.select<AuthModel, bool>(
-      (m) => isRegisterPage ? m.registerStatus.isFetching() : m.loginStatus.isFetching(),
-    );
-
-    ApiError? loginError = context.select<AuthModel, ApiError?>((m) => m.loginStatus.error);
-    ApiError? registerError = context.select<AuthModel, ApiError?>((m) => m.registerStatus.error);
-
     redirectTo = context.read<AuthModel>().redirectToRoute;
     print("redirectTo $redirectTo");
     RegExp regExp = RegExp(r"app\/([a-fA-F0-9-]{36})");
@@ -133,185 +115,8 @@ class _AuthPageFormState extends State<AuthPageForm> {
             ],
           ),
         ),
-        isRegisterPage ? registerForm(registerError) : loginForm(loginError),
+        isRegisterPage ? RegisterForm() : LoginForm(),
       ],
     );
-  }
-
-  Widget registerForm(ApiError? registerError) {
-    return Form(
-      key: _formKey,
-      child: LenraFlex(
-        spacing: 24,
-        direction: Axis.vertical,
-        crossAxisAlignment: CrossAxisAlignment.center,
-        children: [
-          //------Email------
-          LenraTextFormField(
-            label: 'Your email',
-            hintText: 'Email',
-            onChanged: (String value) {
-              email = value;
-            },
-            onSubmitted: (_) {
-              submit();
-            },
-            size: LenraComponentSize.large,
-            type: LenraTextFormFieldType.email,
-            autofillHints: [AutofillHints.email],
-          ),
-          //------Password------
-          LenraTextFormField(
-            label: 'Your password',
-            description: "8 characters, 1 uppercase, 1 lowercase and 1 special character.",
-            obscure: _hidePassword,
-            hintText: 'Password',
-            onSubmitted: (_) {
-              submit();
-            },
-            onChanged: (String value) {
-              password = value;
-            },
-            size: LenraComponentSize.large,
-            type: LenraTextFormFieldType.password,
-            autofillHints: [AutofillHints.newPassword],
-            onSuffixPressed: () {
-              setState(() {
-                _hidePassword = !_hidePassword;
-              });
-            },
-          ),
-          if (registerError != null) Error(registerError),
-          LoadingButton(
-            text: "Create my account",
-            onPressed: () {
-              submit();
-            },
-            loading: isLogging,
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget loginForm(ApiError? loginError) {
-    var themeData = LenraThemeData();
-    return Form(
-      key: _formKey,
-      child: LenraFlex(
-        spacing: 24,
-        direction: Axis.vertical,
-        crossAxisAlignment: CrossAxisAlignment.center,
-        children: [
-          LenraTextFormField(
-            label: 'Your email',
-            hintText: 'Email',
-            onChanged: (String value) {
-              email = value;
-            },
-            onSubmitted: (_) {
-              submit();
-            },
-            type: LenraTextFormFieldType.email,
-            size: LenraComponentSize.large,
-            autofillHints: [AutofillHints.email],
-          ),
-          //------Password------
-          LenraTextFormField(
-            label: 'Your password',
-            obscure: _hidePassword,
-            hintText: 'Password',
-            onChanged: (String value) {
-              password = value;
-            },
-            onSubmitted: (_) {
-              submit();
-            },
-            type: LenraTextFormFieldType.password,
-            size: LenraComponentSize.large,
-            autofillHints: [AutofillHints.password],
-            onSuffixPressed: () {
-              setState(() {
-                _hidePassword = !_hidePassword;
-              });
-            },
-          ),
-          LenraFlex(
-            fillParent: true,
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: [
-              LenraCheckbox(
-                value: keep,
-                onPressed: (value) {
-                  setState(() {
-                    keep = value!;
-                  });
-                },
-              ),
-              LenraText(
-                text: "Keep me logged in",
-                style: themeData.lenraTextThemeData.disabledBodyText,
-              ),
-            ],
-          ),
-          if (loginError != null) Error(loginError),
-          LenraFlex(
-            direction: Axis.vertical,
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: [
-              TextButton(
-                onPressed: !isLogging
-                    ? () {
-                        submit();
-                      }
-                    : null,
-                child: LenraText(text: "Login", style: TextStyle(color: Colors.white)),
-                style: ButtonStyle(
-                  backgroundColor: MaterialStateProperty.resolveWith((states) {
-                    if (states.contains(MaterialState.disabled)) {
-                      return themeData.lenraColorThemeData.primaryBackgroundDisabledColor;
-                    } else if (states.contains(MaterialState.hovered)) {
-                      return themeData.lenraColorThemeData.primaryBackgroundHoverColor;
-                    }
-                    return themeData.lenraColorThemeData.primaryBackgroundColor;
-                  }),
-                  shape: MaterialStateProperty.all(
-                    RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(6),
-                    ),
-                  ),
-                  fixedSize: MaterialStateProperty.all(Size(136, 36)),
-                ),
-              ),
-              Container(
-                padding: EdgeInsets.only(top: 10.0),
-                child: RichText(
-                  text: TextSpan(
-                    text: "I forgot my password",
-                    style: themeData.lenraTextThemeData.blueBodyText,
-                    recognizer: TapGestureRecognizer()
-                      ..onTap = () {
-                        CommonNavigator.go(context, CommonNavigator.lostPassword);
-                      },
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ],
-      ),
-    );
-  }
-
-  void submit() {
-    if (_formKey.currentState!.validate()) {
-      AuthModel authModel = context.read<AuthModel>();
-      final future = isRegisterPage ? authModel.register(email, password) : authModel.login(email, password, keep);
-      future.then((_) {
-        GoRouter.of(context).go(context.read<AuthModel>().redirectToRoute ?? '/');
-      }).catchError((error) {
-        logger.warning(error);
-      });
-    }
   }
 }
